+++ cli/cli/Contants.cs
@@ -29,6 +29,11 @@ public static class Constants
 	public const string PLATFORM_STAGING = "https://staging.api.beamable.com";
 	public const string PLATFORM_PRODUCTION = "https://api.beamable.com";
 	public const string DEFAULT_PLATFORM = PLATFORM_PRODUCTION;
+
+	public const string DOCKER_REGISTRY_DEV = "https://dev-ecr.beamable.com";
+	public const string DOCKER_REGISTRY_STAGING = "https://staging-ecr.beamable.com";
+	public const string DOCKER_REGISTRY_PRODUCTION = "https://ecr.beamable.com";
+
 	public const string CONFIG_CID = "cid";
 	public const string CONFIG_PID = "pid";
 	public const string CONFIG_DIR = "dir";